refac(zk): separate vanishing related from circuit/expression/evaluator/**
