+++ tachyon/zk/plonk/circuit/expressions/evaluator/test/BUILD.bazel
@@ -4,7 +4,10 @@ tachyon_cc_library(
     name = "evaluator_test",
     testonly = True,
     hdrs = ["evaluator_test.h"],
-    visibility = ["//tachyon/zk/plonk/circuit/expressions/evaluator:__pkg__"],
+    visibility = [
+        "//tachyon/zk/plonk/circuit/expressions/evaluator:__pkg__",
+        "//tachyon/zk/plonk/vanishing:__pkg__",
+    ],
     deps = [
         "//tachyon/math/finite_fields/test:gf7",
         "//tachyon/math/polynomials/univariate:univariate_evaluations",
