+++ tachyon/zk/plonk/keys/BUILD.bazel
@@ -7,8 +7,8 @@ tachyon_cc_library(
     hdrs = ["proving_key.h"],
     deps = [
         ":verifying_key",
-        "//tachyon/zk/plonk/circuit/expressions/evaluator:cl_evaluator",
         "//tachyon/zk/plonk/permutation:permutation_proving_key",
+        "//tachyon/zk/plonk/vanishing:cl_evaluator",
     ],
 )
