+++ tachyon/zk/plonk/vanishing/calculation.h
@@ -4,8 +4,8 @@
 // can be found in the LICENSE-MIT.halo2 and the LICENCE-APACHE.halo2
 // file.
 
-#ifndef TACHYON_ZK_PLONK_CIRCUIT_EXPRESSIONS_EVALUATOR_CALCULATION_H_
-#define TACHYON_ZK_PLONK_CIRCUIT_EXPRESSIONS_EVALUATOR_CALCULATION_H_
+#ifndef TACHYON_ZK_PLONK_VANISHING_CALCULATION_H_
+#define TACHYON_ZK_PLONK_VANISHING_CALCULATION_H_
 
 #include <string>
 #include <utility>
@@ -15,8 +15,8 @@
 
 #include "tachyon/base/logging.h"
 #include "tachyon/export.h"
-#include "tachyon/zk/plonk/circuit/expressions/evaluator/value_source.h"
-#include "tachyon/zk/plonk/circuit/expressions/evaluator/value_source_data.h"
+#include "tachyon/zk/plonk/vanishing/value_source.h"
+#include "tachyon/zk/plonk/vanishing/value_source_data.h"
 
 namespace tachyon::zk {
 
@@ -224,4 +224,4 @@ class TACHYON_EXPORT Calculation {
 
 }  // namespace tachyon::zk
 
-#endif  // TACHYON_ZK_PLONK_CIRCUIT_EXPRESSIONS_EVALUATOR_CALCULATION_H_
+#endif  // TACHYON_ZK_PLONK_VANISHING_CALCULATION_H_
