+++ tachyon/zk/plonk/vanishing/value_source_data.h
@@ -4,8 +4,8 @@
 // can be found in the LICENSE-MIT.halo2 and the LICENCE-APACHE.halo2
 // file.
 
-#ifndef TACHYON_ZK_PLONK_CIRCUIT_EXPRESSIONS_EVALUATOR_VALUE_SOURCE_DATA_H_
-#define TACHYON_ZK_PLONK_CIRCUIT_EXPRESSIONS_EVALUATOR_VALUE_SOURCE_DATA_H_
+#ifndef TACHYON_ZK_PLONK_VANISHING_VALUE_SOURCE_DATA_H_
+#define TACHYON_ZK_PLONK_VANISHING_VALUE_SOURCE_DATA_H_
 
 #include <stddef.h>
 
@@ -33,4 +33,4 @@ struct ValueSourceData {
 
 }  // namespace tachyon::zk
 
-#endif  // TACHYON_ZK_PLONK_CIRCUIT_EXPRESSIONS_EVALUATOR_VALUE_SOURCE_DATA_H_
+#endif  // TACHYON_ZK_PLONK_VANISHING_VALUE_SOURCE_DATA_H_
